[feat] #47 Lineに表示するviewに関するロジックをapp/helpers以下に追加
